+++ apps/backend/src/app/api/v1/teams/crud.tsx
@@ -1,12 +1,14 @@
+import { ensureTeamMembershipExist } from "@/lib/db-checks";
 import { isTeamSystemPermission, teamSystemPermissionStringToDBType } from "@/lib/permissions";
 import { sendWebhooks } from "@/lib/webhooks";
 import { prismaClient } from "@/prisma-client";
 import { createCrudHandlers } from "@/route-handlers/crud-handler";
+import { getIdFromUserIdOrMe } from "@/route-handlers/utils";
 import { Prisma } from "@prisma/client";
 import { KnownErrors } from "@stackframe/stack-shared";
 import { teamsCrud } from "@stackframe/stack-shared/dist/interface/crud/teams";
-import { yupObject, yupString } from "@stackframe/stack-shared/dist/schema-fields";
-import { StatusError } from "@stackframe/stack-shared/dist/utils/errors";
+import { userIdOrMeSchema, yupObject, yupString } from "@stackframe/stack-shared/dist/schema-fields";
+import { StatusError, throwErr } from "@stackframe/stack-shared/dist/utils/errors";
 
 
 export function teamPrismaToCrud(prisma: Prisma.TeamGetPayload<{}>) {
@@ -20,11 +22,11 @@ export function teamPrismaToCrud(prisma: Prisma.TeamGetPayload<{}>) {
 
 export const teamsCrudHandlers = createCrudHandlers(teamsCrud, {
   querySchema: yupObject({
-    user_id: yupString().optional(),
+    user_id: userIdOrMeSchema.optional(),
     add_current_user: yupString().oneOf(["true", "false"]).optional(),
   }),
   paramsSchema: yupObject({
-    team_id: yupString().required(),
+    team_id: yupString().uuid().required(),
   }),
   onCreate: async ({ query, auth, data }) => {
     const db = await prismaClient.$transaction(async (tx) => {
@@ -84,18 +86,30 @@ export const teamsCrudHandlers = createCrudHandlers(teamsCrud, {
     return teamPrismaToCrud(db);
   },
   onRead: async ({ params, auth }) => {
-    const db = await prismaClient.team.findUnique({
-      where: {
-        projectId_teamId: {
+    const db = await prismaClient.$transaction(async (tx) => {
+      if (auth.type === 'client') {
+        await ensureTeamMembershipExist(tx, {
           projectId: auth.project.id,
           teamId: params.team_id,
-        },
-      },
-    });
+          userId: auth.user?.id ?? throwErr("Client must be logged in to read a team"),
+        });
+      }
 
-    if (!db) {
-      throw new KnownErrors.TeamNotFound(params.team_id);
-    }
+      const db = await prismaClient.team.findUnique({
+        where: {
+          projectId_teamId: {
+            projectId: auth.project.id,
+            teamId: params.team_id,
+          },
+        },
+      });
+
+      if (!db) {
+        throw new KnownErrors.TeamNotFound(params.team_id);
+      }
+
+      return db;
+    });
 
     return teamPrismaToCrud(db);
   },
@@ -134,17 +148,10 @@ export const teamsCrudHandlers = createCrudHandlers(teamsCrud, {
     });
   },
   onList: async ({ query, auth }) => {
-    if (auth.type === 'client') {
-      if (query.user_id !== 'me' && query.user_id !== auth.user?.id) {
-        throw new StatusError(StatusError.Forbidden, "You are only allowed to access your own teams with the client access token.");
-      }
+    const userId = getIdFromUserIdOrMe(query.user_id, auth.user);
+    if (auth.type === 'client' && userId !== auth.user?.id) {
+      throw new StatusError(StatusError.Forbidden, 'Client can only list teams for their own user. user_id must be either "me" or the ID of the current user');
     }
-
-    if (query.user_id === 'me' && !auth.user) {
-      throw new KnownErrors.CannotGetOwnUserWithoutUser();
-    }
-
-    let userId = query.user_id === 'me' ? auth.user?.id : query.user_id;
 
     const db = await prismaClient.team.findMany({
       where: {