+++ apps/dashboard/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @stackframe/stack-dashboard
 
+## 2.5.6
+
+### Patch Changes
+
+- Various bugfixes
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.6
+  - @stackframe/stack@2.5.6
+  - @stackframe/stack-emails@2.5.6
+
 ## 2.5.5
 
 ### Patch Changes