+++ examples/docs-examples/CHANGELOG.md
@@ -1,5 +1,14 @@
 # demo-app
 
+## 2.5.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.6
+  - @stackframe/stack@2.5.6
+  - @stackframe/stack-ui@2.5.6
+
 ## 2.5.5
 
 ### Patch Changes