+++ examples/e-commerce/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @stackframe/e-commerce-demo
 
+## 2.5.6
+
+### Patch Changes
+
+- @stackframe/stack@2.5.6
+
 ## 2.5.5
 
 ### Patch Changes