+++ examples/partial-prerendering/CHANGELOG.md
@@ -1,5 +1,12 @@
 # partial-prerendering
 
+## 2.5.6
+
+### Patch Changes
+
+- Various bugfixes
+  - @stackframe/stack@2.5.6
+
 ## 2.5.5
 
 ### Patch Changes