+++ packages/stack-ui/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @stackframe/stack-ui
 
+## 2.5.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.6
+
 ## 2.5.5
 
 ### Patch Changes